### 2020-07-22 #2504 @ketch
> @ketch can you check that missing DOI? ☝️

Sure -- that reference is a book.  The suggested DOI refers to a review of the book, so it isn't correct.  As far as I know, the book doesn't have a DOI.
