### 2020-07-23 #2504 @ChrisRackauckas
Thanks for invitation, but I do not think I will be able to turn this around in an adequate time. Good luck to anyone else suddenly building an online course too 😱.
