### 2020-07-23 #2504 @mancellin
Thank you for the invitation. Unfortunately, I don't think I'll be able to make it either.
