### 2020-07-23 #2504 @diehlpk
@ketch Can you please add the countries no abbreviations to all the affiliations?
