### 2020-07-23 #2514 @whedon
[REVIEW]: RK-Opt: A package for the design of numerical ODE solvers
**Submitting author:** @ketch (<a href="http://orcid.org/0000-0002-1212-126X">David Ketcheson</a>)
**Repository:** <a href="https://github.com/ketch/RK-Opt" target ="_blank">https://github.com/ketch/RK-Opt</a>
**Version:** v1.0.3
