### 2020-08-29 ketch/RK-Opt#53 @gardner48
License file formatting
Consider reformatting the LICENSE file so GitHub can automatically identify it as the BSD 3-clause license with [licensee](https://github.com/licensee/licensee).
[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)
