### 2020-08-30 bbf5649
Update to latest standard version of BSD-3 license.
Closes #53.
