### 2020-08-30 ketch/RK-Opt#54 @gardner48
Documentation comments and suggestions
[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)

I think the documentation looks excellent. The only overarching comment I have is that the function documentation style is a little inconsistent. Some functions have a brief description and a bulleted list of inputs/outputs with a description for each item while others just have the brief description which may not cover all the inputs/outputs for the function.

Below is a list of minor comments and typos or style inconsistencies I noticed in reading through the docs. I've linked to the corresponding line in the function comments since there does not seem to be an easy way to do that for the rst files.

* The style for variable names in the input/output lists is some times italic and some times not inconsistent e.g., `rk_obj`.
* `k` is missing from the list of inputs in the rst docs.
https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/nonlinear_constraints.m#L15
* "Methods" to "methods." https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/oc_albrecht.m#L4
* Changing "RKMs" to "RK methods" would be clearer https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/oc_butcher.m#L4
* The meaning of `solveoderconditions=1` is not clear here it would be good to have a note to see the documentation for `rk_opt`
* "initialize" to "initialized" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/rk_opt.m#L25
* "controll" to "control" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/rk_opt.m#L39
* In the rst docs, "tored" to "stored" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/unpack_rk.m#L6
* `writeField` does not have a description https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RK-coeff-opt/write_field.m#L2
* ". different" to "for different" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/multi_R_opt.m#L6
* In the rst docs this line and the next run together  https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/multi_R_opt.m#L9
* Not sure if this indentation was intentional https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/multi_R_opt.m#L21
* Note the inputs `p` and `q` are arrays of polynomial coefficients for the numerator and denominator respectively https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/radimpfast.m#L2
* `R` is missing from `outputs` in `Rkp`, `Rkp_dw`, `Rkp_imp`, and `Rkp_imp_dw` e.g., https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rkp.m#L10
* I think this was meant to be a bulleted list like `Rkp` and `Rkp_imp` https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rkp_dw.m#L7
* missing `)` after first `F`, missing backslash on second `beta`, and I believe `t` is intended to be `\tilde{}` https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rkp_dw.m#L13
* `tbeta` is missing from the list of outputs https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rkp_imp_dw.m#L2
* I think this was meant to be a bulleted list like `Rkp` and `Rkp_imp` https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rkp_imp_dw.m#L7
* Fill in the general stability function https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rskp.m#L19
* In the rst docs "epends" to "Depends" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/am_radius-opt/Rskp.m#L21
* I really like the example calls included in the `opt_poly_bisect` documentation especially the second example that refers to a figure in a publication. Where relevant/applicable, it would be great to have similar examples in the docs for other functions.
* It would be nice if the references in the `spectrum` documentation linked to the articles (or a full citation). https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/polyopt/spectrum.m#L11
* Add "`c` should be a column vector of length `m`" https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RKtools/am_radius.m#L8
* "10^-{10})" to "10^{-10})." https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RKtools/am_radius.m#L10
* In the rst docs the equations after "such that" are missing https://github.com/ketch/RK-Opt/blob/bbf5649260be0874a48f5465b2dfe135ea2eba1c/RKtools/am_radius.m#L15
* `rk_stabfun` uses `rk.A` and `rk.b` and `plotstabreg` accesses `rk.Ahat`, `rk.bhat`, and `rk.chat`. It would be good to document what fields can/should be in an `rk` variable.
* I noticed the contents of the `dwrk_opt` and `low_storage` directories are not documented are these pieces of RK-Opt that are in development?
