### 2020-08-31 ketch/RK-Opt#55 @gardner48
CONTRIBUTING.md
[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)

In addition to the Contributing section in the documentation it would be good to have a top-level `CONTRIBUTING` file as well so users see a link to it when they open a PR. It would also be good to note if you prefer users create a GitHub issue or use some other means for reporting problems.
