### 2020-08-31 ketch/RK-Opt#56 @gardner48
JOSS Paper Statement of need
[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)

The [Automated design of Runge-Kutta methods](https://github.com/ketch/RK-Opt/blob/master/doc/rk-methods.rst) section in the documentation includes some nice motivational material about the use of additional stages to improve method properties and the intractability of finding optimal solutions for the larger set of coefficients by hand. I think it would be good to include some of this at the start of the "Statement of need" section of the JOSS software paper.
