### 2020-08-31 04b1a0f
Add CONTRIBUTING.md.
Closes #55.
