+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+Contributions to RK-Opt are most welcome!  If you have an idea for a new
+feature, or simply see something that should be fixed/improved, you can:
+
+  - [Raise an issue to let us know](https://github.com/ketch/RK-Opt/issues)
+  - Better yet, fork the RK-Opt repository on Github, implement the change
+    yourself, and [open a pull request](https://github.com/ketch/RK-Opt/pulls)
+
+Ideally, contributions that include new functionality should also include
+documentation and testing, but we can help with that part as needed.
+
+If you're raising an issue to report a bug, please provide enough detail to
+let us reproduce the bug.
