### 2020-08-31 f20899e
Address issues raised in #54 (mainly documentation).
