+++ am_radius-opt/Rskp.m
@@ -4,21 +4,21 @@ function [R,gamma]=Rskp(s,k,p)
 % Finds the optimal contractive k-step, s-stage GLM with order of accuracy p
 % for linear problems
 %
-% Inputs: s = # of stages
-%         k = # of steps
-%         p = order of accuracy
+% Inputs:
+%       * s = # of stages
+%       * k = # of steps
+%       * p = order of accuracy
 %
-% Outputs: 
-%        R = threshold factor
-%        gamma = coefficients of the polynomials
+% Outputs:
+%       * R = threshold factor
+%       * gamma = coefficients of the polynomials
 %         
 %        for k=1, the resulting polynomial is
 %        `\sum_{j=0}^m (1+z/R)^j`
 %
-%        in general, the resulting stability function is
-%        (Fill in)
+%        For details on the general case, see :cite:`2009_monotonicity`.
 %
-%Depends on MATLAB's optimization toolbox for the LP solver
+% This routine requires MATLAB's optimization toolbox for the LP solver.
     
 %=========================================================
 %Initialize
