+++ am_radius-opt/Rkp_dw.m
@@ -4,13 +4,16 @@ function [R,alpha,beta,tbeta]=Rkp_dw(k,p)
 % Finds the optimal SSP k-step explicit LMM with order of accuracy p
 % allowing downwind operators
 %
-% Inputs: k = # of steps
-%         p = order of accuracy
+% Inputs:
+%       * k = # of steps
+%       * p = order of accuracy
 %
-% Outputs: alpha, beta, tbeta = the coefficients of the method
+% Outputs:
+%       * R = the SSP coefficient
+%       * alpha, beta, tbeta = the coefficients of the method
 %
 % The method is given by
-% `u_n = \sum_{j=0}^{k-1} (\alpha[j] + \beta[j] F(u_{n-k+j} + tbeta[j] tF(u_{n-k+j}))`
+% `u_n = \sum_{j=0}^{k-1} (\alpha[j] + \beta[j] F(u_{n-k+j}) + tbeta[j] tF(u_{n-k+j}))`
 % where tF(u) is the negated downwind operator.
 %
 % Depends on MATLAB's optimization toolbox for the LP solver
