### 2020-08-31 ketch/RK-Opt#54 @ketch
@gardner48 Thank you so much for the careful review and the many helpful suggestions!

I've addressed most of these in f20899ed319d39b233fc75be5eaeda710bf582e2; the online docs do not reflect the updates yet but will update later today.  Below I've responded to the issues that I didn't fix, or that were a bit more complex.

>The meaning of `solveoderconditions=1` is not clear here it would be good to have a note to see the documentation for `rk_opt`

I'm not sure where "here" refers to in this comment -- as far as I can see, this parameter only appears in `rk_opt.m` and `order_conditions.m`.

> missing backslash on second beta, and I believe t is intended to be \tilde{}

The problem here is making clear the correspondence between the variable tbeta and the coefficient in the formula.  It is written in the way it is so there is no confusion.  Suggestions for improvement are welcome, but rendering the symbol as it should be mathematically would obscure the connection to the MATLAB variable.

>I noticed the contents of the dwrk_opt and low_storage directories are not documented are these pieces of RK-Opt that are in development?

Yes, this is research code that was never really polished but could still be of use.  I've added a README for each of these directories with a short explanation and a warning.
