### 2020-08-31 cf6c782
Add discussion of many-stage RK methods to JOSS paper.
As suggested in #56.
