### 2020-08-31 ketch/RK-Opt#56 @ketch
Thanks -- I forgot that I had written that, but it does fit very well in the paper.
