### 2020-08-31 72ca0a5
More minor improvements to the docs.
