+++ README.md
@@ -1,7 +1,7 @@
 
 # RK-Opt: A Package for the Design of Numerical ODE solvers
 
-See the full documentation [here](http://numerics.kaust.edu.sa/RK-Opt).
+See the full documentation [here](https://rk-opt.readthedocs.io/en/latest/).
 
 RK-Opt is a collection of MATLAB codes for designing optimized numerical ODE solvers.
 The main emphasis is on Runge-Kutta methods, but some routines deal with other classes of methods.
@@ -17,7 +17,7 @@ It includes the following sub-packages:
    Includes capabilities for both multistep and multistage methods.
  - **RKtools**: A collection of routines for analyzing or computing various
    properties of Runge-Kutta methods.  For a much more extensive package along these
-   lines, see [NodePy](http://numerics.kaust.edu.sa/nodepy).
+   lines, see [NodePy](http://nodepy.readthedocs.io/en/latest/).
 
 A common workflow for designing Runge-Kutta methods is to use **polyopt** to find an
 appropriate stability function and then **RK-coeff-opt** to determine the Runge-Kutta
