### 2020-08-31 ketch/RK-Opt#54 @ketch
I've also just switched to using readthedocs.io as the primary location for the docs, as I think it will be easier to keep them in sync that way.  I'm still keeping a mirror up on numerics.kaust.edu.sa since I think there are links to that floating around the internet.
