### 2020-08-31 ketch/RK-Opt#54 @ketch
I think that now all issues raised here have been resolved.
