### 2020-08-31 #2514 @emconsta
@ketch I am using a commercial Matlab license and do not have access to all toolboxes. In addition to required toolboxes,
 I believe parpool is in 'Parallel Computing Toolbox'. I think that should be listed as a requirement or optional requirement since it's not critical.
