### 2020-08-31 ketch/RK-Opt#54 @gardner48
> I'm not sure where "here" refers to in this comment -- as far as I can see, this parameter only appears in `rk_opt.m` and `order_conditions.m`.

Oops, I forgot to include a pointer to the relevant comment. I was referring to the documentation for `order_conditions.m` https://github.com/ketch/RK-Opt/blob/3b4bd7e816b89c884b54278b511f64a5a4d4c51c/RK-coeff-opt/order_conditions.m#L4
Just looking at the documentation for this function it was not clear that `solverorderconditions` is an input to `rk_opt` which then calls this function if `solveorderconditions=1`.
