### 2020-08-31 ketch/RK-Opt#54 @gardner48
> The problem here is making clear the correspondence between the variable tbeta and the coefficient in the formula. It is written in the way it is so there is no confusion. Suggestions for improvement are welcome, but rendering the symbol as it should be mathematically would obscure the connection to the MATLAB variable.

Thanks for the explanation. Just looking at the formula in the documentation I'd thought this was a typo since the other variables are rendered as mathematical symbols but aligning with the MATLAB variable name makes sense.
