### 2020-08-31 7492bd9
Specify where solveorderconditions is used.
As suggested in #54.
