### 2020-08-31 ketch/RK-Opt#54 @ketch
>Just looking at the documentation for this function it was not clear that solverorderconditions is an input to rk_opt which then calls this function if solveorderconditions=1.

Thanks; I've added that to the function documentation.
