### 2020-08-31 #2514 @ketch
@emconsta You're right; thanks for pointing that out.  I've added a note to the README in 67e0916.
