### 2020-09-28 #2514 @diehlpk
@ketch I checked the references and please update

1. Murray, L. M. (2013). Bayesian state-space modelling on high-performance hardware usingLibBi. (add the arxiv.org information)
2. Taylor, S. (1982). Financial returns modelled by the product of two stochastic processes, astudy of daily sugar prices 1961-79,1. (Add the journal)
3. Brown, T. R. (2020). Approximating posterior predictive distributions by averaging outputfrom many particle filters. (Add archiv.org information)
