### 2020-10-05 ketch/RK-Opt#57 @gardner48
Example inputs
[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)

In `multi_am_radius_2_rk_opt.m` there is a missing `fgets` that causes an error when using the example input file given in the comments (or the empty line before the first separation line in the example input needs to be removed).
```
--- a/examples/multi_am_radius_2_rk_opt.m
+++ b/examples/multi_am_radius_2_rk_opt.m
@@ -157,6 +156,9 @@ function [s,k,p,coeffs_monomial]= read_data(read_fid,class)
 %
 % Read number of stage and order of the method

+    % Read white line
+    tline = fgets(read_fid);
+
     % Read separation line
     tline = fgets(read_fid);
```
There also a missing `;` earlier in the file.
```
--- a/examples/multi_am_radius_2_rk_opt.m
+++ b/examples/multi_am_radius_2_rk_opt.m
@@ -64,7 +64,7 @@ for i_stab_poly = 1:number_stab_poly
         % The order of accuracy is set to one to re-use completely the
         % functions implemented in the RK-coeff-opt to impose just the
         % linear order conditions.
-        h_p = 1
+        h_p = 1;

         % Polynomial coefficients indicies
         poly_coeff_ind = 2:s;
```
It would be helpful to include the example input in the directory so a user can run the example without having to create the input file first.

For `multi_stab_fun_2_rk_opt.m` what is parsed by `read_data` didn't seem to match what is described in the comments and what is used in the `multi_stab_fun_2_rk_opt` function. I made the following changes
```
--- a/examples/multi_stab_fun_2_rk_opt.m
+++ b/examples/multi_stab_fun_2_rk_opt.m
@@ -29,7 +29,7 @@ function multi_rk = multi_stab_fun_2_rk_opt(input_file_name,class,objective,vara
 for i_stab_poly = 1:number_stab_poly

     % Read number of stage and order of accuracy of the RK scheme
-    [s,p,fp,d]= read_data(read_fid);
+    [s,p,fp,poly_coeff_val]= read_data(read_fid);

     % Set tall tree numbers (indices) and tall tree values that will be
     % used to enforce the stability coefficients. Indeed, when "s" is
@@ -52,14 +52,15 @@ for i_stab_poly = 1:number_stab_poly
     end

     % Polynomial coefficients indices
-    poly_coeff_ind = p+1:s;
-
-    % Polynomial coefficients values
-    poly_coeff_val = d(7+p+1:length(d));
+    if (fp > 0)
+        poly_coeff_ind = p+fp:s;
+    else
+        poly_coeff_ind = [];
+    end

     % Call to rk_opt
     rk = rk_opt(s,p,class,objective,'poly_coeff_ind',poly_coeff_ind,...
-                'poly_coeff_val',poly_coeff_val,varargin{:})
+                'poly_coeff_val',poly_coeff_val,varargin{:});

 end

@@ -105,8 +106,8 @@ end

 % =========================================================================

-function [s,p,fp,d]= read_data(read_fid)
-%function [s,p,fp]= read_data(read_fid)
+function [s,p,fp,poly_coeff]= read_data(read_fid)
+%function [s,p,fp,poly_coeff]= read_data(read_fid)
 %
 % Read number of stage and order of accuracy of the method

@@ -127,6 +128,9 @@ p = floor(d(2));
 % the order of accuracy of the scheme for nonlinear problems.
 fp = floor(d(3));

+% Polynomial coefficients values
+poly_coeff = d(4:length(d));
+
 end
 % =========================================================================
```
and ran the example with with the following input file:
```
stability poly
1

stages order free params. poly coeff

5      4     1            1 1 0.5  0.166666666666667 0.041666666666667 0.009615384615385
```
I'm sure it this is the intended usage but it seemed to work as expected. I think it would be helpful in include a sample input file for this example as well.

For both of the above examples it would also be helpful to include some comments about the expected output or maybe have an example output file for a user to compare against.

I also tried out the examples given in the `opt_poly_bisect` documentation as well as other setups from KA12. I've included the commands I ran below in case anyone else also wants to try reproducing plots from KA12. I'm not sure if it was intentional but I noticed the gap size is fixed at `20` when using the 'gap' option in `spectrum`.
```
%% Doc example
lam = spectrum('realaxis',500);
s = 10; p = 2;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'chebyshev')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 3 top
lam = spectrum('realaxis',6400);
s = 20; p = 4;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'chebyshev')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 3 bottom
lam = spectrum('realaxis',6400);
s = 20; p = 10;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'chebyshev')
plotstabreg_func(poly_coeff,[1],[-140 1 -10 10])

%% KA12 - Fig 4 left
lam = spectrum('imagaxis',3200);
s = 7; p = 1;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'rotated chebyshev')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 4 right
lam = spectrum('imagaxis',3200);
s = 7; p = 4;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'rotated chebyshev')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 6 left
lam = spectrum('disk',500);
s = 8; p = 3;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'monomial')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 6 right
lam = spectrum('disk',500);
s = 15; p = 4;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'monomial')
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 8 (size of gap is hard coded to 20)
lam = spectrum('gap',500);
s = 6; p = 1;
[h,poly_coeff] = opt_poly_bisect(lam,s,p,'chebyshev')
plotstabreg_func(poly_coeff,[1],[-50 2 -10 10])

%% KA12 - Fig 10 top
lam_func = @(kappa) spectrum('rectangle',100,kappa,1)
[h,poly_coeff] = opt_poly_bisect(lam,10,1,'chebyshev','lam_func',lam_func)
plotstabreg_func(poly_coeff,[1])

%% KA12 - Fig 10 bottom
lam_func = @(kappa) spectrum('rectangle',100,kappa,10)
[h,poly_coeff] = opt_poly_bisect(lam,20,1,'chebyshev','lam_func',lam_func)
```
