### 2020-10-15 ketch/RK-Opt#57 @ketch
Regarding the examples directory: thanks for going over this so carefully and debugging this code!  In fact, I had forgotten about this code completely; it was written by someone who left the project long ago and hasn't been touched since.  My feeling is that these examples are needlessly complicated and the scripts are useful only in a rather specific situation, so I will remove them.  I think simple examples that just call rk_opt once (like those available in the documentation) are more helpful.

The list of inputs to reproduce the KA12 paper is great; I will add that to the project.
