### 2020-10-18 4a19a0d
Remove examples directory.
This directory contained old code that hasn't been maintained
or used.  The smaller examples given in the other subdirectories
are preferred.

Also add example inputs in polyopt/.

Thanks to @gardner48.

Closes #57.
