### 2020-10-18 ketch/RK-Opt#59 @debdeepbh
Documentation
Here are some minor suggestions for the documentation.

* In https://rk-opt.readthedocs.io/en/latest/started.html section "Obtaining RK-Opt":
In the line "add the subdirectory RK-Opt/RKtools", it would be nice to mention how it is done (`addpath`).

* In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html Line 5:
Extra tick around the code snippet `results_rkopt = runtests('test_rkopt.m'); table(results_rkopt)`.

* Same comment for test code snippet in https://rk-opt.readthedocs.io/en/latest/polyopt.html

* In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html section "rk_opt" in the fist "Note:" block:
Capitalize the first letter. Also it would be nice add a reference pointing toward the said numerical experiment.

* In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html
Please follow a single convention to specify the input and output variables associated with all the functions. For example, the description of the input variables for the functions `oc_butcher`, `unpack_lsrk` are scattered across the page. It would be helpful for the reader to find those descriptions near the function description, even if they are redundant. Maybe stick to the style of categorizing "Input" and "Output" variables followed in https://rk-opt.readthedocs.io/en/latest/am_radius-opt.html

* Same comment for various function in https://rk-opt.readthedocs.io/en/latest/RKtools.html and some of the function in https://rk-opt.readthedocs.io/en/latest/am_radius-opt.html

* In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html section "oc_ksrk", the warning could be in a block instead of `...warning`.

* In https://rk-opt.readthedocs.io/en/latest/polyopt.html section "opt_poly_bisect":
In "solvers:" spelling for "licenses"
In "lam_func:" the line "see Figure 10 of the CAMCoS paper": should add direct link to reference.

* In https://rk-opt.readthedocs.io/en/latest/RKtools.html, last line of math mode:
A period to complete the sentence.

[JOSS Review 2514](https://github.com/openjournals/joss-reviews/issues/2514)
