### 2020-10-19 5d38bfd
Minor improvements to the documentation.
Thanks to @debdeepbh.
Closes #59.
