+++ doc/refs.bib
@@ -38,3 +38,11 @@
 }
 
 
+@article{2017_msrk,
+	Author = {Bresten, Christopher and Gottlieb, Sigal and Grant, Zachary and Higgs, Daniel and Ketcheson, David I. and N{\'e}meth, Adrian},
+	Journal = {Math. of Comp.},
+	Pages = {747-769},
+	Title = {Strong Stability Preserving Multistep {R}unge-{K}utta Methods},
+	Volume = {86},
+	Year = {2017}}
+
