### 2020-10-19 ketch/RK-Opt#59 @ketch
Thanks for the careful checking of the docs.

> * In https://rk-opt.readthedocs.io/en/latest/started.html section "Obtaining RK-Opt":
>   In the line "add the subdirectory RK-Opt/RKtools", it would be nice to mention how it is done (`addpath`).

There are multiple ways to do this; I've added a link to the MATLAB documentation.

> * In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html Line 5:
>   Extra tick around the code snippet `results_rkopt = runtests('test_rkopt.m'); table(results_rkopt)`.
> * Same comment for test code snippet in https://rk-opt.readthedocs.io/en/latest/polyopt.html

Fixed.

> * In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html section "rk_opt" in the fist "Note:" block:
>   Capitalize the first letter. Also it would be nice add a reference pointing toward the said numerical experiment.

Fixed.  The experiments in question are not recorded; that's why the note is here.

> * In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html
>   Please follow a single convention to specify the input and output variables associated with all the functions. For example, the description of the input variables for the functions `oc_butcher`, `unpack_lsrk` are scattered across the page. It would be helpful for the reader to find those descriptions near the function description, even if they are redundant. Maybe stick to the style of categorizing "Input" and "Output" variables followed in https://rk-opt.readthedocs.io/en/latest/am_radius-opt.html

These functions are not intended to be called by the user, so we haven't spent much effort on their documentation.  Unfortunately, in MATLAB there doesn't seem to be a standard way of indicating that a function is "private".
Nevertheless, I've tried to improve the documentation of these functions.

> * Same comment for various function in https://rk-opt.readthedocs.io/en/latest/RKtools.html and some of the function in https://rk-opt.readthedocs.io/en/latest/am_radius-opt.html
> * In https://rk-opt.readthedocs.io/en/latest/RK-coeff-opt.html section "oc_ksrk", the warning could be in a block instead of `...warning`.

Fixed.

> * In https://rk-opt.readthedocs.io/en/latest/polyopt.html section "opt_poly_bisect":
>   In "solvers:" spelling for "licenses"

That's the correct British spelling.  Some people involved in the project use British spelling and others use American, and it doesn't seem worthwhile to enforce one strictly.

>   In "lam_func:" the line "see Figure 10 of the CAMCoS paper": should add direct link to reference.

Fixed.

> * In https://rk-opt.readthedocs.io/en/latest/RKtools.html, last line of math mode:
>   A period to complete the sentence.

Fixed.
