### 2020-10-19 #2514 @diehlpk
Since we quite close that the reviewers finished their review, I will do some editorial tasks on the paper,

1. Affiliations: Please add the country
1. Would it be possible to add the grant number for the AFOSR grant? Same for the DOE Computational Science Graduate Fellowship?
1. Some citations use abbreviations (minority) for the journal name and some use the full name. Could you make this consistent?
