### 2020-10-20 d50b047
Minor changes to paper.
Fix capitalization in bibliography.
Add countries in affiliations.
Add AFOSR grant number.
