+++ paper.bib
@@ -25,7 +25,7 @@
 
 @inproceedings{parsani-eccomas,
 	Author = {Parsani, Matteo and Ketcheson, David I. and Deconinck, W.},
-	Booktitle = {Proceedings of the European Congress on Computational Methods in Applied Sciences and Engineering},
+	Booktitle = {{Proceedings of the European Congress on Computational Methods in Applied Sciences and Engineering}},
 	Month = {September},
 	Title = {Explicit {R}unge-{K}utta schemes for the spectral difference method applied to wave propagation problems},
 	Year = {2012}}
@@ -33,7 +33,7 @@
 
 @inproceedings{Parsani_finnish,
 	Author = {Parsani, Matteo and Ketcheson, David I.},
-	Booktitle = {Proceedings of the 11th Finnish Mechanics Days},
+	Booktitle = {{Proceedings of the 11th Finnish Mechanics Days}},
 	Month = {November},
 	Pages = {1--5},
 	Title = {Optimized low-order explicit {R}unge-{K}utta schemes for high-order spectral difference method},
@@ -75,8 +75,8 @@
 
 @inproceedings{2018_wso,
 	Author = {Ketcheson, David I. and Seibold, Benjamin and Shirokoff, David and Zhou, Dong},
-	Booktitle = {Spectral and High Order Methods for Partial Differential Equations ICOSAHOM 2018},
-	Editor = {Sherwin, Spencer et. al.w},
+	Booktitle = {{Spectral and High Order Methods for Partial Differential Equations (ICOSAHOM) 2018}},
+	Editor = {Sherwin, Spencer J and Moxey, David and Peir{\'o}, Joaquim and Vincent, Peter E and Schwab, Christoph}
 	Title = {DIRK Schemes with High Weak Stage Order},
 	Year = {2018}}
 
@@ -118,7 +118,7 @@
 @inproceedings{reynoso2017strong,
   title={Strong stability preserving sixth order two-derivative {R}unge--{K}utta methods},
   author={Reynoso, Gustavo Franco and Gottlieb, Sigal and Grant, Zachary J},
-  booktitle={AIP Conference Proceedings},
+  booktitle={{AIP Conference Proceedings}},
   volume={1863},
   number={1},
   pages={560068},
@@ -166,7 +166,7 @@
 	Month = feb,
 	Number = {2},
 	Pages = {373--392},
-	Title = {{Optimal implicit strong stability preserving {R}unge--{K}utta methods}},
+	Title = {Optimal implicit strong stability preserving {R}unge--{K}utta methods},
 	Volume = {59},
     doi = {10.1016/j.apnum.2008.03.034},
 	Year = {2009}}
@@ -210,7 +210,7 @@
 
 @article{2017_msrk,
 	Author = {Bresten, Christopher and Gottlieb, Sigal and Grant, Zachary and Higgs, Daniel and Ketcheson, David I. and N{\'e}meth, Adrian},
-	Journal = {Math. of Comp.},
+	Journal = {Mathematics of Computation},
 	Pages = {747-769},
 	Title = {Strong Stability Preserving Multistep {R}unge-{K}utta Methods},
 	Volume = {86},
@@ -283,7 +283,7 @@
  Author = {I. {Higueras} and T. {Rold\'an}},
  Title = {{New third order low-storage SSP explicit Runge-Kutta methods}},
  FJournal = {{Journal of Scientific Computing}},
- Journal = {{J. Sci. Comput.}},
+ Journal = {{Journal of Scientific Computing}},
  ISSN = {0885-7474; 1573-7691/e},
  Volume = {79},
  Number = {3},
