### 2020-10-20 #2514 @ketch
> 1. Affiliations: Please add the country

Done.

> 2. Would it be possible to add the grant number for the AFOSR grant? Same for the DOE Computational Science Graduate Fellowship?

There isn't a grant number for the CSGF.  I've added the AFOSR number.

> 3. Some citations use abbreviations (minority) for the journal name and some use the full name. Could you make this consistent?

Done.

> 4. For some references the capitalization seems to be strange, e.g. In Proceedings of the european congress on computational methods in applied sciences and engineering. Could you check these? Adding `{}` to the title or journal name will keep the capitalization.

Done; I think I caught them all.
