+++ paper.bib
@@ -76,7 +76,7 @@
 @inproceedings{2018_wso,
 	Author = {Ketcheson, David I. and Seibold, Benjamin and Shirokoff, David and Zhou, Dong},
 	Booktitle = {{Spectral and High Order Methods for Partial Differential Equations (ICOSAHOM) 2018}},
-	Editor = {Sherwin, Spencer J and Moxey, David and Peir{\'o}, Joaquim and Vincent, Peter E and Schwab, Christoph}
+	Editor = {Sherwin, Spencer J and Moxey, David and Peir{\'o}, Joaquim and Vincent, Peter E and Schwab, Christoph},
 	Title = {DIRK Schemes with High Weak Stage Order},
 	Year = {2018}}
 
