### 2020-10-22 #2514 @diehlpk
@ketch

- [ ] Please generate a new release including all the updates from the review
- [ ] Please post the new release number here
- [ ] Please upload the release and obtain a DOI (Zenodo or FigShare). Note that the title and the authors have to match with the paper.
- [ ] Post the DOI here
