### 2020-10-27 #2514 @ketch
@diehlpk I've tried to do what you asked.  The release version number is 1.0.1 and the DOI is 10.5281/zenodo.4138077.  I don't see any author information or any way to set the author information on Zenodo.
