### 2020-10-27 #2514 @ranocha
I've submitted a PR to RK-Opt to fix that: https://github.com/ketch/RK-Opt/pull/60
