### 2020-10-27 #2514 @ketch
Thanks, @ranocha .  I think everything is in order now.  Here is the corrected information:

Release: 1.0.2
DOI: 10.5281/zenodo.4138949
