### 2020-10-27 #2514 @ranocha
I've submitted a PR to make the title on Zenodo match the title of the paper: https://github.com/ketch/RK-Opt/pull/62.

The order of authors looks to me identical on Zenodo and the paper (David I. Ketcheson;  Matteo Parsani;  Zachary J. Grant;  Aron J. Ahmadia;  Hendrik Ranocha).
