### 2020-10-27 #2514 @diehlpk
Ok, perfect. Once I have the new release and the DOI. I can accept the paper. If you like to avoid having a new release, you can log in to Zendo and change the title there.
