### 2020-10-28 #2514 @ketch
I couldn't find a way to change the title on Zenodo.  Actually, it looks correct there already.  But in case it's needed, I made another release.

New release: 1.0.3
DOI: 10.5281/zenodo.4146740
