### 2020-10-28 #2514 @diehlpk
@ketch

![image](https://user-images.githubusercontent.com/3922435/97459718-0fc70280-190a-11eb-8d8d-ed40d2136f10.png)

The title is still `ketch/RK-OPT` but should be the same title as in the paper.
