### 2020-10-28 #2514 @diehlpk
@ketch

![image](https://user-images.githubusercontent.com/3922435/97460802-17d37200-190b-11eb-95e8-26a5c2164371.png)

There is the edit button and you should be able to replace `ketch/RK-OPT` with the paper title there without doing a new release.
