### 2020-10-29 #2514 @ketch
@diehlpk Thanks, but I can't see that.  Here is what I see:

![Screen Shot 2020-10-29 at 9 34 41 AM](https://user-images.githubusercontent.com/697763/97533899-22a10c00-19ca-11eb-833d-ebda6c233661.png)

Can you tell me how to get to the page with the "edit" and "New version" buttons?
