### 2020-10-29 #2514 @ketch
Okay, I think I finally found it and fixed it:

![Screen Shot 2020-10-29 at 9 38 54 AM](https://user-images.githubusercontent.com/697763/97534193-a2c77180-19ca-11eb-92c1-c6addd210ea0.png)

Although on the other page I screenshotted above, it still shows exactly the same thing as before.
