### 2020-10-30 #2514 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
